fastapi/requirements.txt
# fastapi/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-google-genai
python-multipart